Moved search test models into their own app
