+++ wagtail/tests/search/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'wagtail.tests.search.apps.WagtailSearchTestsAppConfig'
