+++ wagtail/tests/search/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class WagtailSearchTestsAppConfig(AppConfig):
+    name = 'wagtail.tests.search'
+    label = 'searchtests'
+    verbose_name = "Wagtail search tests"
